+++ client/src/hooks/MCP/index.ts
@@ -3,3 +3,4 @@ export * from './useMCPConnectionStatus';
 export * from './useMCPSelect';
 export * from './useVisibleTools';
 export { useMCPServerManager } from './useMCPServerManager';
+export { useRemoveMCPTool } from './useRemoveMCPTool';
